EVA-873 Updated variant start and end from int to long

+++ variation-commons-core/src/main/java/uk/ac/ebi/eva/commons/core/models/IVariant.java
@@ -28,9 +28,9 @@ public interface IVariant {
 
     String getChromosome();
 
-    int getStart();
+    long getStart();
 
-    int getEnd();
+    long getEnd();
 
     int getLength();
